+++ templates/element/paginator.php
@@ -21,23 +21,23 @@ if (!$this->Paginator->hasPage(2, null)) {
 
 <nav class="d-print-none mt-4">
     <ul class="pagination d-none d-lg-flex justify-content-center m-0">
-        <?= $this->Paginator->prev('', ['icon' => 'caret-left']) ?>
+        <?= $this->Paginator->prev() ?>
         <?= $this->Paginator->numbers() ?>
-        <?= $this->Paginator->next('', ['icon' => 'caret-right']) ?>
+        <?= $this->Paginator->next() ?>
     </ul>
     <ul class="pagination d-lg-none justify-content-center m-0">
         <?php
         if ($this->Paginator->hasPrev() && $this->Paginator->hasNext()) {
-            echo $this->Paginator->prev('', ['icon' => 'caret-left']);
+            echo $this->Paginator->prev();
             echo $this->Html->li(
-                $this->Html->span(__d('me_tools', 'Page {0}', $this->Paginator->current()), ['class' => 'page-link']),
+                $this->Html->link(__d('me_tools', 'Page {0}', $this->Paginator->current()), '#', ['class' => 'page-link']),
                 ['class' => 'page-item']
             );
-            echo $this->Paginator->next('', ['icon' => 'caret-right']);
+            echo $this->Paginator->next();
         } elseif (!$this->Paginator->hasPrev()) {
-            echo $this->Paginator->next(__d('me_tools', 'Next'), ['icon' => 'caret-right', 'icon-align' => 'right']);
+            echo $this->Paginator->next(__d('me_tools', 'Next'));
         } else {
-            echo $this->Paginator->prev(__d('me_tools', 'Previous'), ['icon' => 'caret-left']);
+            echo $this->Paginator->prev(__d('me_tools', 'Previous'));
         }
         ?>
     </ul>