+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # 2.x branch
 ## 2.25 branch
+### 2.25.7
+* sorting links (and their icons) are now fully managed by `PaginatorHelper` using the `$_defaultConfig` property.
+  Removed css rules;
+* "next" and "prev" links (and their icons) are now fully managed by `PaginatorHelper` using the `$_defaultConfig` property;
+* fixed `FormHelper::createInline()`.
+
 ### 2.25.6
 * css rules for ascending/descending sort links no longer depend on tables;
 * fixed margins for `paginator.php` element;